[WIP] cpubits: initial crate
As discussed in #824, adds a crate with the intent of it providing
heuristics for selecting whether 32-bit or 64-bit backends have optimal
codegen for a given target, with optional overrides.

The intended use of this crate is in `build-dependencies`, where it can
emit a `cfg` attribute (e.g. `--cfg cpubits="64"`) if one hasn't been
explicitly specified already, and all gating on 32-bit vs 64-bit
backends can simply use the `cfg` attribute.